remove unused component, api
