+++ frontend/src/pages/Home/index.tsx
@@ -2,7 +2,6 @@ import ExamplePosts from "./components/ExamplePosts"
 import Demo from "./components/Demo"
 import Features from "./components/Features"
 import Hero from "./components/Hero"
-import EmotionAnalyzer from "./components/Emotion-test/EmotionAnalyzer"
 
 const Home = () => (
   <div>
@@ -10,7 +9,6 @@ const Home = () => (
     <Features />
     <Demo />
     <ExamplePosts />
-    <EmotionAnalyzer />
   </div>
 )
